backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.3.1
joblib==1.5.3
mssql-django==1.7
numpy==2.4.4
pandas==3.0.2
pyodbc==5.3.0
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
gunicorn==23.0.0
